member 등록, 수정, 삭제 시 password 필요(password 는 passwordEncoder 로 암호화되어 저장)

+++ src/main/java/com/sparta/schedule/service/MemberService.java
@@ -1,25 +1,32 @@
 package com.sparta.schedule.service;
 
+import com.sparta.schedule.config.PasswordEncoder;
 import com.sparta.schedule.dto.member.MemberRequestDto;
 import com.sparta.schedule.dto.member.MemberResponseDto;
 import com.sparta.schedule.entity.Member;
 import com.sparta.schedule.repository.MemberRepository;
 import jakarta.transaction.Transactional;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class MemberService {
 
     private final MemberRepository memberRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    @Autowired
     public MemberService(MemberRepository memberRepository) {
         this.memberRepository = memberRepository;
+        this.passwordEncoder = new PasswordEncoder();
     }
 
     // POST
     public MemberResponseDto createMember(MemberRequestDto createMemberRequestDto) {
         Member member = new Member(createMemberRequestDto);
-        Member savedMember = memberRepository.save(member);
-        return new MemberResponseDto(savedMember);
+        member.setPassword(passwordEncoder.encode(createMemberRequestDto.getPassword()));
+        memberRepository.save(member);
+        return new MemberResponseDto(member);
     }
 
     // GET
@@ -32,14 +39,23 @@ public class MemberService {
     @Transactional
     public MemberResponseDto updateMember(Long userKey, MemberRequestDto updateMemberRequestDto) {
         Member member = findByKey(userKey);
+
+        if(!passwordEncoder.matches(updateMemberRequestDto.getPassword(), member.getPassword())) {
+            throw new RuntimeException("비밀번호가 일치하지 않습니다.");
+        }
+
         member.update(updateMemberRequestDto);
+        member.setPassword(passwordEncoder.encode(updateMemberRequestDto.getPassword()));
         MemberResponseDto memberResponseDto = new MemberResponseDto(member);
         return memberResponseDto;
     }
 
     // DELETE
-    public Long deleteMember(Long userKey) {
+    public Long deleteMember(Long userKey, MemberRequestDto deleteMemberRequestDto) {
         Member member = findByKey(userKey);
+        if(!passwordEncoder.matches(deleteMemberRequestDto.getPassword(), member.getPassword())) {
+            throw new RuntimeException("비밀번호가 일치하지 않습니다.");
+        }
         memberRepository.delete(member);
         return userKey;
     }
